drivers: scmi-msg: refine local header files
Change scmi-msg drivers to define/declare protocol resources in
protocol specific local file instead of from local common.h file.
