+++ core/drivers/scmi-msg/base.c
@@ -11,6 +11,7 @@
 #include <trace.h>
 #include <util.h>
 
+#include "base.h"
 #include "common.h"
 
 static bool message_id_is_supported(unsigned int message_id);
